requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pyzbar==0.1.9
python-multipart==0.0.9
sqlalchemy==2.0.29
pydantic==1.10.14
qrcode==7.4.2
pytest==8.1.1
httpx==0.27.2
